# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: QR Studio API
  version: 1.0.0
  description: |
    QR Studio REST API for QR code management, analytics, and automation.
    
    ## Authentication
    
    Two authentication methods are supported:
    
    1. **Session-based**: Use NextAuth session cookies (for web app)
    2. **API Key**: Use Bearer token in Authorization header (for integrations)
    
    ## Rate Limiting
    
    - Free: 100 requests/hour
    - Pro: 1,000 requests/hour
    - Business: 10,000 requests/hour
    
    ## Base URL
    
    - Production: `https://qr-studio.com`
    - Development: `http://localhost:3000`
    
  contact:
    name: API Support
    email: support@qr-studio.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://qr-studio.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key authentication
    
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Session-based authentication
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum: [qr.created, qr.updated, qr.deleted, qr.scanned, scan.created]
        isActive:
          type: boolean
        secret:
          type: string
        createdAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
  - SessionAuth: []
paths:
...
  /api/webhooks:
    get:
      summary: List webhooks
      tags:
        - Webhooks
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'

    post:
      summary: Create webhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum: [qr.created, qr.updated, qr.deleted, qr.scanned, scan.created]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
...